Fixed GNU warning
git-svn-id: svn://svn.code.sf.net/p/opalvoip/code/ptlib/trunk@26311 023b2edf-31b2-4de3-b41e-bca80c47788f

+++ src/ptlib/common/contain.cxx
@@ -821,7 +821,7 @@ template <typename T> PINDEX p_unsigned2string(T value, T base, char * str)
 #endif
 template <typename T> PINDEX p_signed2string(T value, T base, char * str)
 {
-  if (value >= 0)
+  if ((int)value >= 0)
     return p_unsigned2string<T>(value, base, str);
 
   *str = '-';
